Add PersonalDetailsUtils. Move methods out of OptionsListUtils.
DRY up code a bit

Move createOption method

Fix outdated comment

clean up store file

Move getOptions to own file

Split up utils so that OptionsListUtils onyl have methods that call getOptions()
