+++ src/libs/PersonalDetailsUtils.js
@@ -1,0 +1,60 @@
+import _ from 'underscore';
+import Str from 'expensify-common/lib/str';
+import CONST from '../CONST';
+import md5 from './md5';
+
+/**
+ * Helper method to return a default avatar
+ *
+ * @param {String} [login]
+ * @returns {String}
+ */
+function getDefaultAvatar(login = '') {
+    // There are 8 possible default avatars, so we choose which one this user has based
+    // on a simple hash of their login (which is converted from HEX to INT)
+    const loginHashBucket = (parseInt(md5(login).substring(0, 4), 16) % 8) + 1;
+    return `${CONST.CLOUDFRONT_URL}/images/avatars/avatar_${loginHashBucket}.png`;
+}
+
+/**
+ * Adds expensify SMS domain (@expensify.sms) if login is a phone number and if it's not included yet
+ *
+ * @param {String} login
+ * @return {String}
+ */
+function addSMSDomainIfPhoneNumber(login) {
+    if (Str.isValidPhone(login) && !Str.isValidEmail(login)) {
+        return login + CONST.SMS.DOMAIN;
+    }
+    return login;
+}
+
+/**
+ * Returns the personal details for an array of logins
+ *
+ * @param {Array} logins
+ * @param {Object} personalDetails
+ * @returns {Array}
+ */
+function getPersonalDetailsForLogins(logins, personalDetails) {
+    return _.map(logins, (login) => {
+        let personalDetail = personalDetails[login];
+
+        if (!personalDetail) {
+            personalDetail = {
+                login,
+                displayName: login,
+                avatar: getDefaultAvatar(login),
+            };
+        }
+
+        return personalDetail;
+    });
+}
+
+
+export {
+    getDefaultAvatar,
+    addSMSDomainIfPhoneNumber,
+    getPersonalDetailsForLogins,
+};
